kodi: replace with kodi sdp version
