+++ kodi/commands.py
@@ -1,39 +1,42 @@
-#   'commandname':  {'method': command for JSON-RPC,        'set': writa,   'get': use retval, param': names of params,                   'values': param values,ID=playerid,VAL=data,tuple="eval"       'bounds': optional bounds, tuple is range, list is valid items}
+#!/usr/bin/env python3
+# vim: set encoding=utf-8 tabstop=4 softtabstop=4 shiftwidth=4 expandtab
+
+""" commands for dev kodi """
 
 commands = {
-    'update':         {'special': True, 'set': True},  # call-only item to trigger update of all status fields
-    'player':         {'special': True, 'set': False},  # stores player_id
-    'state':          {'special': True, 'set': False},  # stores kodi state
-    'media':          {'special': True, 'set': False},  # stores media type
-    'title':          {'special': True, 'set': False},  # stores media title
-    'streams':        {'special': True, 'set': False},  # stores available audio streams
-    'subtitles':      {'special': True, 'set': False},  # stores available subtitles
-    'macro':          {'special': True, 'set': True},  # triggers macro evaluation
+    'info': {
+        'player':           {'read': True,  'write': False, 'opcode': 'player',                    'reply_pattern': '*', 'item_type': 'num',  'dev_datatype': 'raw', 'params': None, 'item_attrs': {'read_group_levels': 0}},
+        'state':            {'read': True,  'write': False, 'opcode': 'media',                     'reply_pattern': '*', 'item_type': 'str',  'dev_datatype': 'raw', 'params': None, 'item_attrs': {'read_group_levels': 0}},
+        'media':            {'read': True,  'write': False, 'opcode': 'media',                     'reply_pattern': '*', 'item_type': 'str',  'dev_datatype': 'raw', 'params': None, 'item_attrs': {'read_group_levels': 0}},
+        'title':            {'read': True,  'write': False, 'opcode': 'title',                     'reply_pattern': '*', 'item_type': 'str',  'dev_datatype': 'raw', 'params': None, 'item_attrs': {'read_group_levels': 0}},
+        'streams':          {'read': True,  'write': False, 'opcode': 'streams',                   'reply_pattern': '*', 'item_type': 'list', 'dev_datatype': 'raw', 'params': None, 'item_attrs': {'read_group_levels': 0}},
+        'subtitles':        {'read': True,  'write': False, 'opcode': 'subtitles',                 'reply_pattern': '*', 'item_type': 'list', 'dev_datatype': 'raw', 'params': None, 'item_attrs': {'read_group_levels': 0}},
+        'macro':            {'read': True,  'write': True,  'opcode': 'macro',                     'reply_pattern': '*', 'item_type': 'bool', 'dev_datatype': 'raw', 'params': None, 'item_attrs': {'read_group_levels': 0}},
+    },
+    'status': {
+        'update':           {'read': False, 'write': True,  'opcode': 'update',                    'reply_pattern': '*', 'item_type': 'bool', 'dev_datatype': 'raw', 'params': None},
+        'ping':             {'read': True,  'write': False, 'opcode': 'JSONRPC.Ping',              'reply_pattern': '*', 'item_type': 'bool', 'dev_datatype': 'raw', 'params': None},
+        'get_status_au':    {'read': True,  'write': False, 'opcode': 'Application.GetProperties', 'reply_pattern': '*', 'item_type': 'bool', 'dev_datatype': 'raw', 'params': {'properties': ['volume', 'muted']}},
 
-    'ping':           {'method': 'JSONRPC.Ping',              'set': False,   'get': True,    'params': None,                               'values': None,                                                'bounds': None},
-    'get_status_au':  {'method': 'Application.GetProperties', 'set': False,   'get': True,    'params': ['properties'],                     'values': [['volume', 'muted']],                               'bounds': None},
-    'get_players':    {'method': 'Player.GetPlayers',         'set': False,   'get': True,    'params': None,                               'values': None,                                                'bounds': None},
-    'get_actplayer':  {'method': 'Player.GetActivePlayers',   'set': False,   'get': True,    'params': None,                               'values': None,                                                'bounds': None},
-    'get_status_play':{'method': 'Player.GetProperties',      'set': False,   'get': True,    'params': ['playerid', 'properties'],         'values': ['ID', ['type', 'speed', 'time', 'percentage', 'totaltime', 'position', 'currentaudiostream', 'audiostreams', 'subtitleenabled', 'currentsubtitle', 'subtitles', 'currentvideostream', 'videostreams']], 'bounds': None},
-    'get_item':       {'method': 'Player.GetItem',            'set': False,   'get': True,    'params': ['playerid', 'properties'],         'values': ['ID', ['title', 'artist']],                         'bounds': None},
-    'get_favourites': {'method': 'Favourites.GetFavourites',  'set': False,   'get': True,    'params': ['properties'],                     'values': [['window', 'path', 'thumbnail', 'windowparameter']],'bounds': None},
+        'get_players':      {'read': True,  'write': False, 'opcode': 'Player.GetPlayers',         'reply_pattern': '*', 'item_type': 'bool', 'dev_datatype': 'raw', 'params': None},
+        'get_actplayer':    {'read': True,  'write': False, 'opcode': 'Player.GetActivePlayers',   'reply_pattern': '*', 'item_type': 'bool', 'dev_datatype': 'raw', 'params': None},
+        'get_status_play':  {'read': True,  'write': False, 'opcode': 'Player.GetProperties',      'reply_pattern': '*', 'item_type': 'bool', 'dev_datatype': 'raw', 'params': {'playerid': '{ID}', 'properties': ['type', 'speed', 'time', 'percentage', 'totaltime', 'position', 'currentaudiostream', 'audiostreams', 'subtitleenabled', 'currentsubtitle', 'subtitles', 'currentvideostream', 'videostreams']}},
+        'get_item':         {'read': True,  'write': False, 'opcode': 'Player.GetItem',            'reply_pattern': '*', 'item_type': 'bool', 'dev_datatype': 'raw', 'params': {'playerid': '{ID}', 'properties': ['title', 'artist']}},
+        'get_favourites':   {'read': True,  'write': False, 'opcode': 'Favourites.GetFavourites',  'reply_pattern': '*', 'item_type': 'bool', 'dev_datatype': 'raw', 'params': {'properties': ['window', 'path', 'thumbnail', 'windowparameter']}},
+    },
+    'control': {
+        'playpause':        {'read': True,  'write': True,  'opcode': 'Player.PlayPause',          'reply_pattern': '*', 'item_type': 'bool', 'dev_datatype': 'raw', 'params': {'playerid': '{ID}', 'play': 'toggle'}, 'item_attrs': {'read_group_levels': 0}},
+        'seek':             {'read': True,  'write': True,  'opcode': 'Player.Seek',               'reply_pattern': '*', 'item_type': 'num',  'dev_datatype': 'raw', 'params': {'playerid': '{ID}', 'value': '{VALUE}'}, 'cmd_settings': {'force_min': 0.0, 'force_max': 100.0}, 'item_attrs': {'read_group_levels': 0}},
+        'audio':            {'read': True,  'write': True,  'opcode': 'Player.SetAudioStream',     'reply_pattern': '*', 'item_type': 'foo',  'dev_datatype': 'raw', 'params': {'playerid': '{ID}', 'stream': '{VALUE}'}, 'item_attrs': {'read_group_levels': 0}},
+        'speed':            {'read': True,  'write': True,  'opcode': 'Player.SetSpeed',           'reply_pattern': '*', 'item_type': 'num',  'dev_datatype': 'raw', 'params': {'playerid': '{ID}', 'speed': '{VALUE}'}, 'cmd_settings': {'valid_list': [-32,-16,-8,-4,-2,-1,1,2,4,8,16,32]}, 'item_attrs': {'read_group_levels': 0}},
+        'subtitle':         {'read': True,  'write': True,  'opcode': 'Player.SetSubtitle',        'reply_pattern': '*', 'item_type': 'foo',  'dev_datatype': 'raw', 'params': {'playerid': '{ID}', 'subtitle': '{VALUE}', 'enable': ('False if "VAL"=="off" else True',)}, 'item_attrs': {'read_group_levels': 0}},
+        'stop':             {'read': True,  'write': True,  'opcode': 'Player.Stop',               'reply_pattern': '*', 'item_type': 'bool', 'dev_datatype': 'raw', 'params': {'playerid': '{ID}'}, 'item_attrs': {'read_group_levels': 0}},
+        'goto':             {'read': True,  'write': True,  'opcode': 'Player.GoTo',               'reply_pattern': '*', 'item_type': 'str',  'dev_datatype': 'raw', 'params': {'playerid': '{ID}', 'to': '{VALUE}'}, 'cmd_settings': {'valid_list': ['previous','next']}, 'item_attrs': {'read_group_levels': 0}},
+        'power':            {'read': True,  'write': True,  'opcode': 'System.Shutdown',           'reply_pattern': '*', 'item_type': 'bool', 'dev_datatype': 'raw', 'params': None, 'item_attrs': {'read_group_levels': 0}},
+        'quit':             {'read': True,  'write': True,  'opcode': 'Application.Quit',          'reply_pattern': '*', 'item_type': 'bool', 'dev_datatype': 'raw', 'params': None, 'item_attrs': {'read_group_levels': 0}},
+        'mute':             {'read': True,  'write': True,  'opcode': 'Application.SetMute',       'reply_pattern': '*', 'item_type': 'bool', 'dev_datatype': 'raw', 'params': {'mute': '{VALUE}'}, 'item_attrs': {'read_group_levels': 0}},
+        'volume':           {'read': True,  'write': True,  'opcode': 'Application.SetVolume',     'reply_pattern': '*', 'item_type': 'num',  'dev_datatype': 'raw', 'params': {'volume': '{VALUE}'}, 'cmd_settings': {'force_min': 0, 'force_max': 100}, 'item_attrs': {'read_group_levels': 0}},
+        'action':           {'read': True,  'write': True,  'opcode': 'Input.ExecuteAction',       'reply_pattern': '*', 'item_type': 'str',  'dev_datatype': 'raw', 'params': {'action': '{VALUE}'}, 'cmd_settings': {'valid_list': ['left','right','up','down','select','back','menu','info','pause','stop','skipnext','skipprevious','fullscreen','aspectratio','stepforward','stepback','osd','showsubtitles','nextsubtitle','cyclesubtitle','audionextlanguage','number1','number2','number3','number4','number5','number6','number7','number8','number9','fastforward','rewind','play','playpause','volumeup','volumedown','mute','enter']}, 'item_attrs': {'read_group_levels': 0}}
+    }
 
-    'playpause':      {'method': 'Player.PlayPause',          'set': True,    'get': False,   'params': ['playerid', 'play'],               'values': ['ID', 'toggle'],                                    'bounds': None},
-    'seek':           {'method': 'Player.Seek',               'set': True,    'get': False,   'params': ['playerid', 'value'],              'values': ['ID', 'VAL'],                                       'bounds': (0.0, 100.0)},
-    'audio':          {'method': 'Player.SetAudioStream',     'set': True,    'get': False,   'params': ['playerid', 'stream'],             'values': ['ID', 'VAL'],                                       'bounds': None},
-    'speed':          {'method': 'Player.SetSpeed',           'set': True,    'get': False,   'params': ['playerid', 'speed'],              'values': ['ID', 'VAL'],                                       'bounds': [-32,-16,-8,-4,-2,-1,1,2,4,8,16,32]},
-    'subtitle':       {'method': 'Player.SetSubtitle',        'set': True,    'get': False,   'params': ['playerid', 'subtitle', 'enable'], 'values': ['ID', 'VAL', ('False if "VAL"=="off" else True',)],    'bounds': None},
-    'stop':           {'method': 'Player.Stop',               'set': True,    'get': False,   'params': ['playerid'],                       'values': ['ID'],                                              'bounds': None},
-    'goto':           {'method': 'Player.GoTo',               'set': True,    'get': False,   'params': ['playerid', 'to'],                 'values': ['ID', 'VAL'],                                       'bounds': ['previous','next']},
-    'power':          {'method': 'System.Shutdown',           'set': True,    'get': False,   'params': None,                               'values': None,                                                'bounds': None},
-    'quit':           {'method': 'Application.Quit',          'set': True,    'get': False,   'params': None,                               'values': None,                                                'bounds': None},
-    'mute':           {'method': 'Application.SetMute',       'set': True,    'get': False,   'params': ['mute'],                           'values': ['VAL'],                                             'bounds': None},
-    'volume':         {'method': 'Application.SetVolume',     'set': True,    'get': False,   'params': ['volume'],                         'values': ['VAL'],                                             'bounds': (0, 100)},
-    'action':         {'method': 'Input.ExecuteAction',       'set': True,    'get': False,   'params': ['action'],                         'values': ['VAL'],                                             'bounds': ['left','right','up','down','select','back','menu','info','pause','stop','skipnext','skipprevious','fullscreen','aspectratio','stepforward','stepback','osd','showsubtitles','nextsubtitle','cyclesubtitle','audionextlanguage','number1','number2','number3','number4','number5','number6','number7','number8','number9','fastforward','rewind','play','playpause','volumeup','volumedown','mute','enter']},
-}
-
-# 'macroname': [ [cmd1, val1], [cmd2, val2], [cmd3, val3], ...]. -- cmd can be one of commands.keys() or 'wait' to wait 'val' seconds
-macros = {
-    'resume': [['action', 'play'], ['wait', 1], ['action', 'select']],
-    'beginning': [['action', 'play'], ['wait', 1], ['action', 'down'], ['action', 'select']]
 }
